hyperv: Use a file as marker to know if component was installed
We were checking if a .exe file was existing to know if a component was
installed; however, that didn't work for networking-hyperv, and it's not
super reliable in case there's a file rename upstream.

So instead, just touch a file when we do the installation, and use it to
know if we did the install or not.